Made client name a required argument, modified cleanup_server
A unique name is now required when creating an instance of the client. The cleanup_server method was changed to only clean endpoints and subs that contain the client's unique name. This allows an application to easily clean up connections that it may have left behind after a crash without affecting other subs/endpoints.